% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{default_bandwidth.nrd}
\alias{default_bandwidth.nrd}
\title{Default function given by function bandwidth.nrd in MASS. No changes were made to this function.}
\usage{
default_bandwidth.nrd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A suitable bandwith.
}
\description{
Default function given by function bandwidth.nrd in MASS. No changes were made to this function.
}
