% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{extract_row_dgRMatrix}
\alias{extract_row_dgRMatrix}
\title{Returns a row of a sparse matrix of class dgRMatrix. Function made by Ben Bolker and Ott Toomet (see https://stackoverflow.com/questions/47997184/)}
\usage{
extract_row_dgRMatrix(m, i = 1)
}
\arguments{
\item{m}{a sparse matrix of class dgRMatrix}

\item{i}{the index of the row to return}
}
\value{
A row (vector) of the sparse matrix
}
\description{
Returns a row of a sparse matrix of class dgRMatrix. Function made by Ben Bolker and Ott Toomet (see https://stackoverflow.com/questions/47997184/)
}
\examples{

# make a toy dgRMatrix
m <- matrix(rpois(12,1), nrow=4)
m <- as(m, "dgRMatrix")
# get first row of this dgRMatrix
extract_row_dgRMatrix(m, 1)
}
