% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{extract_row_lgRMatrix}
\alias{extract_row_lgRMatrix}
\title{Returns a row of a sparse matrix of class lgRMatrix.
Function made by Ben Bolker and Ott Toomet (see https://stackoverflow.com/questions/47997184/)}
\usage{
extract_row_lgRMatrix(m, i = 1)
}
\arguments{
\item{m}{a sparse matrix of class lgRMatrix}

\item{i}{the index of the row to return}
}
\value{
A row (logical vector) of the sparse matrix
}
\description{
Returns a row of a sparse matrix of class lgRMatrix.
Function made by Ben Bolker and Ott Toomet (see https://stackoverflow.com/questions/47997184/)
}
