% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{get.log.p.D_KL}
\alias{get.log.p.D_KL}
\title{Estimates the significance of the observed Kullback-Leibler divergence by comparig to randomizations.}
\usage{
get.log.p.D_KL(T.counts, D_KL.observed, D_KL.randomized,
  output.dir = NULL)
}
\arguments{
\item{T.counts}{The number of cells in which a gene is detected.}

\item{D_KL.observed}{A vector of observed Kullback-Leibler divergences.}

\item{D_KL.randomized}{A matrix of Kullback-Leibler divergences of randomized datasets.}

\item{output.dir}{Optional parameter. Default is NULL. If not NULL, some files will be written to this directory.}
}
\value{
A vector of log10 p values.
}
\description{
Estimates the significance of the observed Kullback-Leibler divergence by comparig to randomizations.
}
