% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{get.parameters.density.estimate}
\alias{get.parameters.density.estimate}
\title{Function that decides most of the parameters that will be during the "Haystack" analysis.}
\usage{
get.parameters.density.estimate(x, y, high.resolution = F)
}
\arguments{
\item{x}{x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)}

\item{y}{y-axis coordinates of cells in a 2D representation}

\item{high.resolution}{Logical: should high resolution be used? Default is FALSE.}
}
\value{
A list containing various parameters to use in the analysis.
}
\description{
Function that decides most of the parameters that will be during the "Haystack" analysis.
}
