% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{get_D_KL}
\alias{get_D_KL}
\title{Calculates the Kullback-Leibler divergence between distributions.}
\usage{
get_D_KL(classes, parameters, reference.prob, pseudo)
}
\arguments{
\item{classes}{A logical vector. Values are T is the gene is expressed in a cell, F is not.}

\item{parameters}{Parameters of the analysis, as set by function 'get_parameters_haystack'}

\item{reference.prob}{A reference distribution to calculate the divergence against.}

\item{pseudo}{A pseudocount, used to avoid log(0) problems.}
}
\value{
A numerical value, the Kullback-Leibler divergence
}
\description{
Calculates the Kullback-Leibler divergence between distributions.
}
