% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_continuous.R
\name{get_D_KL_continuous_2D}
\alias{get_D_KL_continuous_2D}
\title{Calculates the Kullback-Leibler divergence between distributions for the 2-dimensional continuous version of haystack.}
\usage{
get_D_KL_continuous_2D(weights, parameters, reference.prob, pseudo)
}
\arguments{
\item{weights}{A numerical vector with expression values of a gene.}

\item{parameters}{Parameters of the analysis, as set by function 'get_parameters_haystack'}

\item{reference.prob}{A reference distribution to calculate the divergence against.}

\item{pseudo}{A pseudocount, used to avoid log(0) problems.}
}
\value{
A numerical value, the Kullback-Leibler divergence
}
\description{
Calculates the Kullback-Leibler divergence between distributions for the 2-dimensional continuous version of haystack.
}
