% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_spatial.R
\name{get_D_KL_continuous_spatial}
\alias{get_D_KL_continuous_spatial}
\title{Calculates the Kullback-Leibler divergence between distributions for the spatial transcriptomics version of haystack.}
\usage{
get_D_KL_continuous_spatial(
  weights,
  density.contributions,
  reference.prob,
  pseudo = 0
)
}
\arguments{
\item{weights}{A numerical vector with expression values of a gene.}

\item{density.contributions}{A matrix of pairwise density contributions.}

\item{reference.prob}{A reference distribution to calculate the divergence against.}

\item{pseudo}{A pseudocount, used to avoid log(0) problems.}
}
\value{
A numerical value, the Kullback-Leibler divergence
}
\description{
Calculates the Kullback-Leibler divergence between distributions for the spatial transcriptomics version of haystack.
}
