% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_highD.R
\name{get_D_KL_highD}
\alias{get_D_KL_highD}
\title{Calculates the Kullback-Leibler divergence between distributions for the high-dimensional version of haystack().}
\usage{
get_D_KL_highD(classes, density.contributions, reference.prob, pseudo = 0)
}
\arguments{
\item{classes}{A logical vector. Values are T is the gene is expressed in a cell, F is not.}

\item{density.contributions}{A matrix of density contributions of each cell (rows) to each center point (columns).}

\item{reference.prob}{A reference distribution to calculate the divergence against.}

\item{pseudo}{A pseudocount, used to avoid log(0) problems.}
}
\value{
A numerical value, the Kullback-Leibler divergence
}
\description{
Calculates the Kullback-Leibler divergence between distributions for the high-dimensional version of haystack().
}
