% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_highD.R
\name{get_dist_two_sets}
\alias{get_dist_two_sets}
\title{Calculate the pairwise Euclidean distances between the rows of 2 matrices.}
\usage{
get_dist_two_sets(set1, set2)
}
\arguments{
\item{set1}{A numerical matrix.}

\item{set2}{A numerical matrix.}
}
\value{
A matrix of pairwise distances between the rows of 2 matrices.
}
\description{
Calculate the pairwise Euclidean distances between the rows of 2 matrices.
}
