% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_highD.R
\name{get_distance_to_nearest_higher_degree}
\alias{get_distance_to_nearest_higher_degree}
\title{Function for finding for a point in space the distance to the nearest point with higher density. See}
\usage{
get_distance_to_nearest_higher_degree(entries_n = length(degree_counts),
  dist, degree_counts)
}
\arguments{
\item{entries_n}{The number of points.}

\item{dist}{A distance object with the distances between all pairs oof points.}

\item{degree_counts}{A vector with the degree or density of each point.}
}
\value{
A vector with distances.
}
\description{
Function for finding for a point in space the distance to the nearest point with higher density. See
}
