% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_highD.R
\name{get_euclidean_distance}
\alias{get_euclidean_distance}
\title{Calculate the Euclidean distance between x and y.}
\usage{
get_euclidean_distance(x, y)
}
\arguments{
\item{x}{A numerical vector.}

\item{y}{A numerical vector.}
}
\value{
A numerical value, the Euclidean distance.
}
\description{
Calculate the Euclidean distance between x and y.
}
