% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_spatial.R
\name{get_parameters_haystack_spatial}
\alias{get_parameters_haystack_spatial}
\title{Decides the pairwise density contributions between cells or spots, which will be used during the "Haystack" analysis.}
\usage{
get_parameters_haystack_spatial(coordinates, bandwidth.multiplier = 1)
}
\arguments{
\item{coordinates}{Coordinates of cells in spatial transcriptomics data.}

\item{bandwidth.multiplier}{A parameter to adjust the bandwitdh used for estimating the density contributions. Default = 1.}
}
\value{
Pairwise density contributions between all pairs of cell or spots.
}
\description{
Decides the pairwise density contributions between cells or spots, which will be used during the "Haystack" analysis.
}
