% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{get_reference}
\alias{get_reference}
\title{Get reference distribution}
\usage{
get_reference(param, use.advanced.sampling = NULL)
}
\arguments{
\item{param}{Parameters of the analysis, as set by function 'get_parameters_haystack'}

\item{use.advanced.sampling}{If NULL naive sampling is used. If a vector is given (of length = no. of cells) sampling is done according to the values in the vector.}
}
\value{
A list with two components, Q for the reference distribution and pseudo.
}
\description{
Get reference distribution
}
