% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_spatial.R
\name{get_reference_spatial}
\alias{get_reference_spatial}
\title{Get reference distribution for spatial transcriptomics data}
\usage{
get_reference_spatial(pairwise.densities, use.advanced.sampling = NULL)
}
\arguments{
\item{pairwise.densities}{Pairwise density contributions between all pairs of grid points}

\item{use.advanced.sampling}{If NULL naive sampling is used. If a vector is given (of length = no. of cells) sampling is done according to the values in the vector.}
}
\value{
A list with two components, Q for the reference distribution and pseudocounts pseudo.
}
\description{
Get reference distribution for spatial transcriptomics data
}
