% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{get_significant_haystack}
\alias{get_significant_haystack}
\title{Returns the most significant results of the 'haystack' analysis}
\usage{
get_significant_haystack(res.haystack, n = 10, p.value.threshold = NA)
}
\arguments{
\item{res.haystack}{A 'haystack' result variable}

\item{n}{If defined, the top "n" sigificant genes will be returned. Default: 10.}

\item{p.value.threshold}{If defined, genes passing this p-value threshold will be returned.}
}
\value{
A subset of significant genes in the 'haystack' result.
}
\description{
Returns the most significant results of the 'haystack' analysis
}
\examples{
warn("I will add this later")
}
