% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_continuous.R
\name{haystack_continuous_2D}
\alias{haystack_continuous_2D}
\title{The main Haystack function, for 2-dimensional spaces and continuous expression levels.}
\usage{
haystack_continuous_2D(
  x,
  y,
  expression,
  weights.advanced.Q = NULL,
  dir.randomization = NULL,
  randomization.count = 100,
  n.genes.to.randomize = 100
)
}
\arguments{
\item{x}{x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)}

\item{y}{y-axis coordinates of cells in a 2D representation}

\item{expression}{a matrix with expression data of genes (rows) in cells (columns)}

\item{weights.advanced.Q}{(Default: NULL) Optional weights of cells for calculating a weighted distribution of expression.}

\item{dir.randomization}{If NULL, no output is made about the random sampling step. If not NULL, files related to the randomizations are printed to this directory.}

\item{randomization.count}{Number of randomizations to use. Default: 100}

\item{n.genes.to.randomize}{Number of genes to use in randomizations. Default: 100}
}
\value{
An object of class "haystack", including the results of the analysis, and the coordinates of the grid points used to estimate densities.
}
\description{
The main Haystack function, for 2-dimensional spaces and continuous expression levels.
}
\examples{
# I need to add some examples.
# A toy example will be added too.
}
