% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_spatial.R
\name{haystack_continuous_spatial}
\alias{haystack_continuous_spatial}
\title{The main Haystack function for spatial transcriptomics data and continuous expression levels.}
\usage{
haystack_continuous_spatial(
  coordinates,
  expression,
  weights.advanced.Q = NULL,
  dir.randomization = NULL,
  scale = FALSE,
  randomization.count = 100,
  n.genes.to.randomize = 100
)
}
\arguments{
\item{coordinates}{Coordinates of cells in spatial transcriptomics data.}

\item{expression}{a matrix with expression data of genes (rows) in cells or spots (columns)}

\item{weights.advanced.Q}{(Default: NULL) Optional weights of cells for calculating a weighted distribution of expression.}

\item{dir.randomization}{If NULL, no output is made about the random sampling step. If not NULL, files related to the randomizations are printed to this directory.}

\item{scale}{Logical (default=FALSE) indicating whether input coordinates should be scaled to mean 0 and standard deviation 1.}

\item{randomization.count}{Number of randomizations to use. Default: 100}

\item{n.genes.to.randomize}{Number of genes to use in randomizations. Default: 100}
}
\value{
An object of class "haystack", including the results of the analysis.
}
\description{
The main Haystack function for spatial transcriptomics data and continuous expression levels.
}
\examples{
# I need to add some examples.
# A toy example will be added too.
}
