% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{haystack_interface}
\alias{haystack_interface}
<<<<<<< HEAD
\title{Interface to various Haystack functions.}
=======
\title{Title}
>>>>>>> 2782eb24f1dc64e0d7865b60a55a4b24ae78569c
\usage{
haystack_interface(
  coordinates,
  expression = NULL,
  detection = NULL,
  type = NULL,
  ...
)
}
\arguments{
\item{coordinates}{typically >2D coordinates of cells in the PC space (2D is also fine), or 2D/3D spatial coordinates}

\item{expression}{matrix or data.frame with continuous expression values of genes in cells/spots}

\item{detection}{matrix or data.frame with binary (TRUE or FALSE) detection values of genes in cells/spots}

\item{type}{highD, 2D, or spatial, depending on the type of input coordinates}

\item{...}{other parameters to pass on to the haystack functions}
}
\value{
An object of class "haystack", including the results of the analysis, and the coordinates of the grid points used to estimate densities.
}
\description{
<<<<<<< HEAD
Interface to various Haystack functions.
=======
Title
>>>>>>> 2782eb24f1dc64e0d7865b60a55a4b24ae78569c
}
\examples{
# I need to add some examples.
# A toy example will be added too.
}
