% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_clustering.R
\name{hclust_haystack_raw}
\alias{hclust_haystack_raw}
\title{Function for hierarchical clustering of genes according to their distribution on a 2D plot.}
\usage{
hclust_haystack_raw(x, y, detection, genes, method = "ward.D")
}
\arguments{
\item{x}{x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)}

\item{y}{y-axis coordinates of cells in a 2D representation}

\item{detection}{A logical matrix showing which genes (rows) are detected in which cells (columns)}

\item{genes}{A set of genes (of the 'detection' data) which will be clustered.}

\item{method}{The method to use for hierarchical clustering. See '?hclust' for more information. Default: "ward.D".}
}
\value{
An object of class hclust, describing a hierarchical clustering tree.
}
\description{
Function for hierarchical clustering of genes according to their distribution on a 2D plot.
}
