% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{kde2d_faster}
\alias{kde2d_faster}
\title{Based on the MASS kde2d() function, but heavily simplified; it's just tcrossprod() now.}
\usage{
kde2d_faster(dens.x, dens.y)
}
\arguments{
\item{dens.x}{Contribution of all cells to densities of the x-axis grid points.}

\item{dens.y}{Contribution of all cells to densities of the y-axis grid points.}
}
\description{
Based on the MASS kde2d() function, but heavily simplified; it's just tcrossprod() now.
}
