% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{kmeans_haystack}
\alias{kmeans_haystack}
\alias{kmeans_haystack.matrix}
\alias{kmeans_haystack.data.frame}
\title{Function for k-means clustering of genes according to their expression distribution in 2D or multi-dimensional space}
\usage{
kmeans_haystack(x, expression, grid.coordinates, k, ...)

\method{kmeans_haystack}{matrix}(x, expression, grid.coordinates, k, ...)

\method{kmeans_haystack}{data.frame}(x, expression, grid.coordinates, k, ...)
}
\arguments{
\item{x}{a matrix or other object from which coordinates of cells can be extracted.}

\item{expression}{expression matrix.}

\item{grid.coordinates}{coordinates of the grid points.}

\item{k}{number of clusters.}

\item{...}{further parameters passed down to methods.}
}
\description{
Function for k-means clustering of genes according to their expression distribution in 2D or multi-dimensional space
}
