% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{plot_compare_ranks}
\alias{plot_compare_ranks}
\title{plot_compare_ranks}
\usage{
plot_compare_ranks(res1, res2, sort_by = "log.p.vals")
}
\arguments{
\item{res1}{haystack result.}

\item{res2}{haystack result.}

\item{sort_by}{column to sort results (default: log.p.vals).}
}
\description{
plot_compare_ranks
}
