% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{plot_rand_fit}
\alias{plot_rand_fit}
\alias{plot_rand_fit.haystack}
\title{plot_rand_fit}
\usage{
plot_rand_fit(x, type = c("mean", "sd"))

\method{plot_rand_fit}{haystack}(x, type = c("mean", "sd"))
}
\arguments{
\item{x}{haystack object.}

\item{type}{whether to plot mean or sd.}
}
\description{
plot_rand_fit
}
