% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{plot_rand_KLD}
\alias{plot_rand_KLD}
\title{plot_rand_KLD}
\usage{
plot_rand_KLD(x, n = 12, log = TRUE, tail = FALSE)
}
\arguments{
\item{x}{haystack result.}

\item{n}{number of genes from randomization set to plot.}

\item{log}{whether to use log of KLD.}

\item{tail}{whether the genes are chosen from the tail of randomized genes.}
}
\description{
Plots the distribution of randomized KLD for each of the genes, together with
the mean and standard deviation, the 0.95 quantile and the 0.95 quantile from
a normal distribution with mean and standard deviations from the distribution
of KLDs. The logCV is indicated in the subtitle of each plot.
}
