% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_IO.R
\name{read_haystack}
\alias{read_haystack}
\title{Function to read haystack results from file.}
\usage{
read_haystack(file)
}
\arguments{
\item{file}{A file containing 'haystack' results to read}
}
\value{
An object of class "haystack"
}
\description{
Function to read haystack results from file.
}
\examples{
warn("I will add this later")
}
