% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{haystack}
\alias{haystack}
\alias{haystack.matrix}
\alias{haystack.data.frame}
\alias{haystack.Seurat}
\alias{haystack.SingleCellExperiment}
\title{The main Haystack function.}
\usage{
haystack(x, ...)

\method{haystack}{matrix}(x, dim1 = 1, dim2 = 2, method = "highD",
  ...)

\method{haystack}{data.frame}(x, ...)

\method{haystack}{Seurat}(x, assay = "RNA", slot = "data",
  coord = "tsne", cutoff = 1, ...)

\method{haystack}{SingleCellExperiment}(x, assay = "counts",
  coord = "TSNE", cutoff = 1, ...)
}
\arguments{
\item{x}{a matrix or other object from which coordinates of cells can be extracted.}

\item{...}{further paramters passed to haystack_raw().}

\item{dim1}{column index or name of matrix for x-axis coordinates.}

\item{dim2}{column index or name of matrix for y-axis coordinates.}

\item{assay}{name of assay data for Seurat method.}

\item{slot}{name of slot for assay data for Seurat method.}

\item{coord}{name of coordinates slot for specific methods.}

\item{cutoff}{cutoff for detection.}
}
\value{
An object of class "haystack"
}
\description{
The main Haystack function.
}
