% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{haystack_interface}
\alias{haystack_interface}
\alias{haystack_interface.matrix}
\alias{haystack_interface.Seurat}
\title{Interface to various Haystack functions.}
\usage{
haystack_interface(object, ...)

\method{haystack_interface}{matrix}(object = NULL, coordinates = NULL, type = NULL, ...)

\method{haystack_interface}{Seurat}(
  object = NULL,
  reduction = "pca",
  assay = NULL,
  slot = "data",
  type = NULL,
  dims = NULL,
  ...
)
}
\arguments{
\item{object}{a matrix with expression values for genes in cells/spots, or a logical matrix indicating detection values, or a Seurat/SingleCellexperiment object containing expression values.}

\item{...}{other parameters to pass on to the haystack functions.}

\item{coordinates}{a matrix with coordinates of cells in PCA space (>=2 PCs) or other type of embedding (tSNE/UMAP), or 2D/3D spatial coordinates.}

\item{type}{highD, 2D, or spatial, depending on the type of input coordinates.}

\item{reduction}{name for the embeddings to be used when using Seurat or SingleCellExperiment objects.}

\item{assay}{name of the assay to be used when using Seurat or SingleCellExperiment objects.}

\item{slot}{name of the slot to be used when using Seurat objects.}

\item{dims}{dimensions to use.}
}
\value{
An object of class "haystack", including the results of the analysis, and the coordinates of the grid points used to estimate densities.
}
\description{
Interface to various Haystack functions.
}
\examples{
# I need to add some examples.
# A toy example will be added too.
}
