% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{hclust_haystack}
\alias{hclust_haystack}
\alias{hclust_haystack.matrix}
\alias{hclust_haystack.data.frame}
\title{hclust_haystack}
\usage{
hclust_haystack(x, ...)

\method{hclust_haystack}{matrix}(x, dim1 = 1, dim2 = 2, ...)

\method{hclust_haystack}{data.frame}(x, dim1 = 1, dim2 = 2, ...)
}
\arguments{
\item{x}{a matrix or other object from which coordinates of cells can be extracted.}

\item{...}{further paramters passed down to methods.}

\item{dim1}{column index or name of matrix for x-axis coordinates.}

\item{dim2}{column index or name of matrix for y-axis coordinates.}
}
\description{
hclust_haystack
}
