% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_clustering.R
\name{kmeans_haystack}
\alias{kmeans_haystack}
\title{Function for k-means clustering of genes according to their distribution on a 2D plot.}
\usage{
kmeans_haystack(x, y, detection, genes, k, ...)
}
\arguments{
\item{x}{x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)}

\item{y}{y-axis coordinates of cells in a 2D representation}

\item{detection}{A logical matrix showing which gens (rows) are detected in which cells (columns)}

\item{genes}{A set of genes (of the 'detection' data) which will be clustered.}

\item{k}{The number of clusters to return.}

\item{...}{Additional parameters which will be passed on to the kmeans function.}
}
\value{
An object of class kmeans, describing a clustering into 'k' clusters
}
\description{
Function for k-means clustering of genes according to their distribution on a 2D plot.
}
\examples{
warning("I will add this later")
}
