% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_visualization.R
\name{plot_gene_haystack}
\alias{plot_gene_haystack}
\title{Visualizing the detection/expression of a gene in a 2D plot}
\usage{
plot_gene_haystack(x, y, gene, expression, detection = NULL,
  high.resolution = F)
}
\arguments{
\item{x}{x-axis coordinates of cells in a 2D representation (e.g. resulting from PCA or t-SNE)}

\item{y}{y-axis coordinates of cells in a 2D representation}

\item{gene}{name of a gene that is present in the input expression data, or a numerical index}

\item{expression}{a logical/numerical matrix showing detection/expression of genes (rows) in cells (columns)}

\item{detection}{an optional logical matrix showing detection of genes (rows) in cells (columns). If left as NULL, the density distribution of the gene is not plotted.}

\item{high.resolution}{logical (default: FALSE). If set to TRUE, the density plot will be of a higher resolution}
}
\value{
A plot
}
\description{
Visualizing the detection/expression of a gene in a 2D plot
}
\examples{
warning("I will add this later")
}
