% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{show_result_haystack}
\alias{show_result_haystack}
\title{Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.}
\usage{
show_result_haystack(res.haystack, n = NA, p.value.threshold = NA,
  gene = NA)
}
\arguments{
\item{res.haystack}{A 'haystack' result variable}

\item{n}{If defined, the top "n" sigificant genes will be returned. Default: NA, which shows all results.}

\item{p.value.threshold}{If defined, genes passing this p-value threshold will be returned.}

\item{gene}{If defined, the results of this (these) gene(s) will be returned.}
}
\value{
A table with a sorted subset of the 'haystack' result according to input parameters.
}
\description{
Shows the results of the 'haystack' analysis in various ways, sorted by significance. Priority of params is genes > p.value.threshold > n.
}
\examples{
warning("I will add this later")
}
