% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellHaystack-package.R
\docType{package}
\name{singleCellHaystack-package}
\alias{singleCellHaystack}
\alias{singleCellHaystack-package}
\title{singleCellHaystack: Finding surprising needles (=genes) in haystacks (=2D representations of single cell transcriptome data)}
\description{
Haystack finds genes that have a biased expression level in a two-dimensional representation of 
    higher-dimensional transcriptome data, such as single-cell RNA-seq data (scRNA-seq). scRNA-seq data is often 
    represented in 2D plots, such as plots of two principal components, or in t-SNE plots. Many genes are not 
    expressed in all cells in these plots, but rather only in a subset of cells, which might be located in 1 or more
    clusters in the 2D plot. Haystack can be used to find such genes, that have a highly non-random pattern of expression.
}
\keyword{internal}
