% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCellHaystack-package.R
\docType{package}
\name{singleCellHaystack-package}
\alias{singleCellHaystack}
\alias{singleCellHaystack-package}
\title{singleCellHaystack: Finding Needles (=differentially Expressed Genes) in Haystacks (=single Cell Data)}
\description{
`singleCellHaystack` is a package for predicting differentially expressed 
    genes (DEGs) in single cell transcriptome data. It does so without relying on 
    clustering of cells into arbitrary clusters! Single-cell RNA-seq (scRNA-seq) 
    data is often processed to fewer dimensions using Principal Component Analysis 
    (PCA) and represented in 2-dimentional plots (e.g. t-SNE or UMAP plots). 
    `singleCellHaystack` uses Kullback-Leibler Divergence to find genes that are 
    expressed in subsets of cells that are non-randomly positioned in a these 
    multi-dimensional spaces or 2D representations.
}
\author{
\strong{Maintainer}: Alexis Vandenbon \email{alexis.vandenbon@gmail.com}

Authors:
\itemize{
  \item Diego Diez \email{diego10ruiz@gmail.com}
}

}
\keyword{internal}
