% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack_IO.R
\name{write_haystack}
\alias{write_haystack}
\title{Function to write haystack result data to file.}
\usage{
write_haystack(x, file)
}
\arguments{
\item{x}{A 'haystack' result variable}

\item{file}{A file to write to}
}
\description{
Function to write haystack result data to file.
}
\examples{
warn("I will add this later")
}
